+++ ReleaseNotes.md
@@ -19,6 +19,10 @@
 
 * Updates
   * Context:
+    * ArrayDataSource: Updated ArrayDataSource and variants to use Equatable instead of AnyObject, since any equatable object can be used.
+    * CollectionViewArrayDataSource: Updated ArrayDataSource and variants to use Equatable instead of AnyObject, since any equatable object can be used.
+    * TableViewArrayDataSource: Updated ArrayDataSource and variants to use Equatable instead of AnyObject, since any equatable object can be used.
+  * Context:
     * AppContextAwareProtocol: Made generic. Now any Structure or Class can be used as AppContext, allowing any project to have an AppContext created to better suit the project requirements.
     * AppContext: Due to changes in AppContextAwareProtocol, AppContext is no longer essential. Besides that, it was kept inside York, being renamed to FullStackAppContext.
   * Extensions: